+++ eth-crypto/src/main/java/com/xwc1125/chain5j/crypto/TransactionEncoder.java
@@ -25,7 +25,7 @@ public class TransactionEncoder {
     }
 
     public static byte[] signMessage(
-            RawTransaction rawTransaction, byte chainId, Credentials credentials) {
+            RawTransaction rawTransaction, int chainId, Credentials credentials) {
         byte[] encodedTransaction = encode(rawTransaction, chainId);
         Sign.SignatureData signatureData = Sign.signMessage(
                 encodedTransaction, credentials.getEcKeyPair());
@@ -35,8 +35,8 @@ public class TransactionEncoder {
     }
 
     public static Sign.SignatureData createEip155SignatureData(
-            Sign.SignatureData signatureData, byte chainId) {
-        byte v = (byte) (signatureData.getV() + (chainId << 1) + 8);
+            Sign.SignatureData signatureData, int chainId) {
+        int v =  (signatureData.getV() + (chainId << 1) + 8);
 
         return new Sign.SignatureData(
                 v, signatureData.getR(), signatureData.getS());
@@ -46,7 +46,7 @@ public class TransactionEncoder {
         return encode(rawTransaction, null);
     }
 
-    public static byte[] encode(RawTransaction rawTransaction, byte chainId) {
+    public static byte[] encode(RawTransaction rawTransaction, int chainId) {
         Sign.SignatureData signatureData = new Sign.SignatureData(
                 chainId, new byte[] {}, new byte[] {});
         return encode(rawTransaction, signatureData);
